requirements.txt
selenium==4.9.1
requests==2.31.0
Pillow==9.5.0
